Refactor with try catch and notificationReducer

+++ bloglist/client/src/reducers/blogsReducer.js
@@ -1,5 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
 import blogsServices from "../services/blogs";
+import { setNotificationWithTime } from "./notificationReducer";
 
 const blogSlice = createSlice({
   name: "blogs",
@@ -30,21 +31,39 @@ export const initializeBlogs = () => {
 
 export const createBlog = (blog, token) => {
   return async (dispatch) => {
-    const addedBlog = await blogsServices.create(blog, token);
-    dispatch(appendBlog(addedBlog));
+    try {
+      const addedBlog = await blogsServices.create(blog, token);
+      dispatch(appendBlog(addedBlog));
+      dispatch(setNotificationWithTime(`A new blog: ${addedBlog.title} by ${addedBlog.author}`, false, 5000));
+    } catch (error) {
+      dispatch(setNotificationWithTime("Error create new blog", true, 5000));
+    }
   };
 };
 
 export const updateBlog = (blog, token) => {
   return async (dispatch) => {
-    const updatedBlog = await blogsServices.update(blog, token);
-    dispatch(modifyBlog(updatedBlog));
+    try {
+      const updatedBlog = await blogsServices.update(blog, token);
+      dispatch(modifyBlog(updatedBlog));
+    } catch (error) {
+      dispatch(setNotificationWithTime("Error update likes", true, 5000));
+    }
   };
 };
 
-export const deleteBlog = (id, token) => {
+export const deleteBlog = (blog, token) => {
   return async (dispatch) => {
-    const removedBlog = await blogsServices.remove(id, token);
-    dispatch(removeBlog(removedBlog));
+    if (!window.confirm(`Do you really want to remove ${blog.title} by ${blog.author}?`)) return;
+
+    try {
+      const removedBlog = await blogsServices.remove(blog.id, token);
+      dispatch(removeBlog(removedBlog));
+      dispatch(
+        setNotificationWithTime(`The blog ${removedBlog.title} by ${removedBlog.author} has been deleted`, false, 5000)
+      );
+    } catch (error) {
+      dispatch(setNotificationWithTime("Error delete blog", true, 5000));
+    }
   };
 };
